Refactored API js lib into seperate npm module
