+++ package.json
@@ -33,12 +33,13 @@
     "i": "^0.3.5",
     "imports-loader": "^0.7.1",
     "install": "^0.10.1",
-    "jquery": "^3.2.1",
+    "jquery": "^3.3.1",
     "jquery-ui": "^1.12.1",
     "jquery.typewatch": "^3.0.0",
     "js-cookie": "^2.1.4",
     "list.js": "^1.5.0",
     "livefilter": "^2.7.1",
+    "mgnify": "^0.0.1",
     "mocha-phantomjs": "^4.1.0",
     "motion-ui": "^1.2.3",
     "npm": "^5.8.0",
