+++ src/js/modules/analysis.js
@@ -2,7 +2,7 @@ const Backbone = require('backbone');
 const _ = require('underscore');
 const INTERPRO_URL = process.env.INTERPRO_URL;
 const Commons = require('../commons');
-const api = require('../components/api');
+const api = require('mgnify').api;
 const util = require('../util');
 const TaxonomyPieChart = require('../components/charts/taxonomy/taxonomyPie');
 const TaxonomyColumnChart = require('../components/charts/taxonomy/taxonomyColumn');
