+++ src/js/modules/browse.js
@@ -2,7 +2,7 @@ const Backbone = require('backbone');
 const _ = require('underscore');
 const util = require('../util');
 const commons = require('../commons');
-const api = require('../components/api');
+const api = require('mgnify').api;
 const Pagination = require('../components/pagination').Pagination;
 const GenericTable = require('../components/genericTable');
 const Commons = require('../commons');
