+++ src/js/modules/index.js
@@ -1,7 +1,7 @@
 const Backbone = require('backbone');
 const _ = require('underscore');
 const $ = require('jquery');
-const api = require('../components/api');
+const api = require('mgnify').api;
 const ebisearch = require('../components/ebisearch');
 const apiUrl = process.env.API_URL;
 const commons = require('../commons');
