+++ src/js/modules/study.js
@@ -1,6 +1,6 @@
 const Backbone = require('backbone');
 const _ = require('underscore');
-const api = require('../components/api');
+const api = require('mgnify').api;
 const Map = require('../components/map');
 const util = require('../util');
